+++ example_test.go
@@ -90,8 +90,9 @@ func Example() {
 	//     }
 	//   ],
 	//   "log_config": {
-	//     "level": "warn",
-	//     "format": "json"
+	//     "level": "debug",
+	//     "format": "json",
+	//     "verbose": 1
 	//   }
 	// }
 }