+++ CHANGELOG.md
@@ -13,8 +13,10 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ### Changed
 
 - Change several metric names to use `.` instead of `_` for 
-  OpenTelemetry consistency.
-- Update to OpenTelemetry-Go SDK version 0.14.
+  OpenTelemetry consistency. (#43)
+- Update to OpenTelemetry-Go SDK version 0.14. (#41)
+- Removed unnecessary code that reduced batching capability. (#45)
+- Truncate server error messages to 256 bytes. (#46)
 
 ### Removed